# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fix_Smart_CMS API",
    "version": "1.0.3",
    "description": "\n        Comprehensive API documentation for Fix_Smart_CMS (NLC-CMS) - A modern complaint management system.\n        \n        ## Features\n        - **Multi-role Authentication**: Citizens, Ward Officers, Maintenance Teams, Administrators\n        - **Complaint Management**: Complete lifecycle from registration to resolution\n        - **Geographic Routing**: Ward-based automatic complaint assignment\n        - **File Management**: Unified attachment system for all file types\n        - **Real-time Notifications**: Email and in-app notification system\n        - **Analytics & Reporting**: Comprehensive dashboard and reporting capabilities\n        \n        ## Authentication\n        Most endpoints require JWT authentication. Include the token in the Authorization header:\n        `Authorization: Bearer <your-jwt-token>`\n        \n        ## Rate Limiting\n        API requests are rate-limited to prevent abuse. Development environment has higher limits.\n        \n        ## Response Format\n        All API responses follow a consistent format:\n        ```json\n        {\n          \"success\": boolean,\n          \"message\": \"string\",\n          \"data\": object | array | null,\n          \"errors\": array (optional)\n        }\n        ```\n      ",
    "contact": {
      "name": "API Support",
      "email": "api-support@nlc-cms.gov.in",
...
    },
...
    },
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "An error occurred"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "Invalid email format"
                }
              }
            }
          }
        }
      },
...
      "SystemConfig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "config123"
          },
          "key": {
            "type": "string",
            "description": "Configuration key",
            "example": "COMPLAINT_ID_PREFIX"
          },
          "value": {
            "type": "string",
            "description": "Configuration value",
            "example": "KSC"
          },
          "type": {
            "type": "string",
            "description": "Configuration type/category",
            "example": "app"
          },
          "description": {
            "type": "string",
            "description": "Configuration description",
            "example": "Prefix for complaint IDs"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
...
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Authentication required. Please provide a valid token."
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "You don't have permission to access this resource."
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/admin/stats/system": {
      "get": {
        "summary": "Get system-wide statistics",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "System statistics"
          }
        }
      }
    },
...
    "/api/config/stats": {
      "get": {
        "summary": "Get cache statistics",
        "tags": [
          "SystemConfig"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Get system configuration cache statistics (Admin only)",
        "responses": {
          "200": {
            "description": "Cache statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "isInitialized": {
                          "type": "boolean"
                        },
                        "configCount": {
                          "type": "integer"
                        },
                        "lastUpdated": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "refreshInterval": {
                          "type": "integer"
                        },
                        "hasAutoRefresh": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/config/admin": {
      "get": {
        "summary": "Get all system configuration (Admin only)",
        "tags": [
          "SystemConfig"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieve all system configuration values including sensitive data",
        "responses": {
          "200": {
            "description": "Admin configuration retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cached": {
                          "type": "object",
                          "description": "Cached configuration values"
                        },
                        "database": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "key": {
                                "type": "string"
                              },
                              "value": {
                                "type": "string"
                              },
                              "type": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "isActive": {
                                "type": "boolean"
                              },
                              "updatedAt": {
                                "type": "string",
                                "format": "date-time"
                              }
                            }
                          }
                        },
                        "cacheStats": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/config/refresh": {
      "post": {
        "summary": "Refresh configuration cache",
        "tags": [
          "SystemConfig"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Manually refresh the system configuration cache",
        "responses": {
          "200": {
            "description": "Cache refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/config": {
      "post": {
        "summary": "Create new system configuration",
        "tags": [
          "SystemConfig"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemConfigCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Configuration created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SystemConfig"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "409": {
            "description": "Configuration key already exists"
          }
        }
      }
    },
    "/api/config/bulk": {
      "post": {
        "summary": "Bulk update system configurations",
        "tags": [
          "SystemConfig"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemConfigBulkUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk update completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "updated": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SystemConfig"
                          }
                        },
                        "created": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SystemConfig"
                          }
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "key": {
                                "type": "string"
                              },
                              "error": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/config/type/{type}": {
      "get": {
        "summary": "Get configurations by type",
        "tags": [
          "SystemConfig"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "description": "Configuration type",
            "schema": {
              "type": "string",
              "example": "app"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configurations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string"
                        },
                        "configs": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "string"
                          }
                        },
                        "count": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/config/pattern/{pattern}": {
      "get": {
        "summary": "Get configurations by pattern",
        "tags": [
          "SystemConfig"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "pattern",
            "in": "path",
            "required": true,
            "description": "Pattern to match against configuration keys",
            "schema": {
              "type": "string",
              "example": "COMPLAINT_"
            }
          },
          {
            "name": "matchType",
            "in": "query",
            "description": "Type of pattern matching",
            "schema": {
              "type": "string",
              "enum": [
                "startsWith",
                "endsWith",
                "includes"
              ],
              "default": "startsWith"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Configurations retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "pattern": {
                          "type": "string"
                        },
                        "matchType": {
                          "type": "string"
                        },
                        "configs": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "string"
                          }
                        },
                        "count": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
...
  }
}